Vulkan: Split vk_backend functions into multiple files
Attempt to improve the organization of vulkan backend by spliting chuncks of
functions into multiple files.
