+++ Makefile
@@ -1,5 +1,9 @@
-HEADERS = opengl/gl_backend.h opengl/extras.h common/input.h vulkan/vk_backend.h vulkan/vk_window.h common/constants.h common/utils.h
-SOURCES = opengl/gl_backend.c opengl/extras.c common/input.c vulkan/vk_backend.c vulkan/vk_window.c main.c
+HEADERS = opengl/gl_backend.h opengl/extras.h \
+		  vulkan/vk_backend.h vulkan/vk_window.h vulkan/vk_instance.h vulkan/vk_logical_device.h vulkan/vk_types.h \
+		  common/input.h common/constants.h common/utils.h
+SOURCES = opengl/gl_backend.c opengl/extras.c \
+		  vulkan/vk_backend.c vulkan/vk_window.c vulkan/vk_instance.c vulkan/vk_logical_device.c \
+		  common/input.c main.c
 LIBS = -lftgl -lGL -lglfw -lvulkan
 INCLUDES = -I opengl -I vulkan -I common -I /usr/include/freetype2
 CFLAGS := -Wall
